BOT/requirements.txt
gradio
streamlit
mistralai
python-dotenv 

